requirements.txt
flask
flask-sqlalchemy
frozen-flask
python-dotenv
cloudinary
requests